+++ commands/fun/8ball.js
@@ -8,7 +8,7 @@ module.exports = {
   run: async (bot, message, args) => {
 
     const Discord = require("discord.js");
-    let fetch = require("node-fetch");
+    const fetch = require("node-fetch");
 
     let question = args.slice(0).join(" ");
     if (!question || args[0] == "help") {
@@ -29,7 +29,7 @@ module.exports = {
       .addField("Question", question)
       .addField("Answer", result.response)
       .setImage(result.url)
-      .setFooter(`${message.server.name}`, `https://i.imgur.com/BkZY6H8.png`);
+      .setFooter(`${message.server.name}`, `https://i.imgur.com/klY5xCe.png`);
 
     message.reply({ content: "The 8ball is working it's magic! :tada:" }).then((m) => {
       setTimeout(() => {